+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix the bug of panicking when the given source code has CPP lines and space-prefixed lines ([#780]).
 - Fix not pretty-printing multiple signatures in a `SPECIALISE` ([#784]).
 - Fix the bug of not pretty-printing multiple module-level warning messages ([#822])
+- Fix the bug of wrongly converting a `newtype` instance to a `data` one ([#839])
 
 ### Removed
 
@@ -371,6 +372,7 @@ This version is accidentally pushlished, and is the same as 5.3.3.
 [@uhbif19]: https://github.com/uhbif19
 [@toku-sa-n]: https://github.com/toku-sa-n
 
+[#839]: https://github.com/mihaimaruseac/hindent/pull/839
 [#829]: https://github.com/mihaimaruseac/hindent/pull/829
 [#822]: https://github.com/mihaimaruseac/hindent/pull/822
 [#784]: https://github.com/mihaimaruseac/hindent/pull/784